lexgen/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
python-docx==1.1.0
PyMuPDF==1.23.7
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
